+++ packages/authhero/src/management-app.ts
@@ -15,6 +15,7 @@ import { promptsRoutes } from "./routes/management-api/prompts";
 import { registerComponent } from "./middlewares/register-component";
 import { DataAdapters } from "@authhero/adapter-interfaces";
 import { createAuthMiddleware } from "./middlewares/authentication";
+import { emailProviderRoutes } from "./routes/management-api/emails";
 
 export interface CreateAuthParams {
   dataAdapter: DataAdapters;
@@ -36,6 +37,7 @@ export default function create(params: CreateAuthParams) {
   const managementApp = app
     .route("/branding", brandingRoutes)
     // .route("/domains", domainRoutes)
+    .route("/email/providers", emailProviderRoutes)
     .route("/users", userRoutes)
     .route("/keys", keyRoutes)
     .route("/users-by-email", usersByEmailRoutes)