+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,18 @@
 # authhero
 
+## 0.14.0
+
+### Minor Changes
+
+- added email providers and removed tickets
+- Added email providers
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.25.0
+
 ## 0.13.2
 
 ### Patch Changes